// packages/dataTable/src/actionsDropdown/types.ts
import React, { MouseEvent } from 'react';

export type ActionTypes = 'default' | 'danger';

export interface ActionProps {
  testId?: string;
  label: string;
  onClick: (event?: MouseEvent) => void;
  disabled?: {
    isDisabled: boolean;
    reason: string;
  };
  icon?: React.ReactNode;
  key: string;
  type?: ActionTypes;
}